add bsd/osx support proclist()

+++ specialized/proclist.h
@@ -4,6 +4,6 @@
 // POSIX
 #include <sys/types.h>
 
-int proclist(pid_t* list, int max_length);
+int proclist(pid_t* list, size_t max_length);
 
 #endif // PROCLIST_H
